move multi line matcher to pattern
