+++ python/ffi_navigator/pattern.py
@@ -3,6 +3,9 @@ from typing import Optional
 
 import re
 import attr
+import numpy as np
+from bisect import bisect
+
 from .lsp import Range, Position, Location
 
 # We use simple regular expression matching
@@ -90,8 +93,8 @@ def re_matcher(rexpr, fcreate, use_search=False):
     """
     Parameters
     ----------
-    macro_names : list
-        List of macro names to match.
+    rexpr : str
+        A regexp pattern to match.
 
     fcreate : Function (match, path, range) -> result.
 
@@ -116,6 +119,46 @@ def re_matcher(rexpr, fcreate, use_search=False):
                 if item:
                     results.append(item)
         return results
+    return _matcher
+
+
+def re_multi_line_matcher(rexpr, fcreate):
+    """ Matches a pattern spanning multiple lines
+
+    Parameters
+    ----------
+    rexpr : str
+        A regexp pattern to match.
+
+    fcreate : Function (match, path, range) -> result.
+    """
+    rexpr = re.compile(rexpr)
+
+    def _matcher(path, lines):
+        source = "\n".join(lines)
+        matches = list(rexpr.finditer(source))
+        if matches == []:
+            return []
+        line_counts = map(lambda line: len(line)+1, lines) # +1 for newline
+        cumsum = np.cumsum(list(line_counts))
+
+        next_begin = 0
+        result = []
+        # find line num, start and end pos for each match
+        for match in matches:
+            line_num = int(bisect(cumsum[next_begin:], match.start())) + next_begin
+            next_begin = line_num
+            line_num_start = line_num
+            line_num_end = line_num
+            if match.group("key_space"):
+                line_num_end += 1
+            pos_start = match.start() - int(cumsum[line_num_start-1])
+            pos_end = match.end() - int(cumsum[line_num_end-1])
+            rg = Range(Position(line_num_start, pos_start), Position(line_num_end, pos_end))
+            result.append(fcreate(match, path, rg))
+
+        return result
+
     return _matcher
